feat: add b2Bucket to consts and move baseUrl to consts
